+++ packages/undercut/src/push.js
@@ -1,4 +1,3 @@
 export * from "./push/push_core.js";
 export * from "./push/push_operations.js";
-// export * from "./push/push_sources.js";
 export * from "./push/push_targets.js";