+++ packages/undercut/src/push/push_targets.js
@@ -1,4 +1,6 @@
-import { assert } from "../utils/assert.js";
+import { assert, assertFunctor } from "../utils/assert.js";
+import { isUndefined, isFunction } from "../utils/language.js";
+import { initializeObserver } from "../utils/observer.js";
 
 export function createPushTarget() {
 	return {
@@ -18,12 +20,10 @@ export function createPushTarget() {
 	};
 }
 
-/**
- * @param {Function} consumer
- * @param {Function} [finalizer]
- * @returns {Observer}
- */
-export function* toConsumer(consumer, finalizer) {
+function* consumerTarget(consumer, finalizer) {
+	assertFunctor(consumer, `consumer`);
+	assert(isUndefined(finalizer) || isFunction(finalizer), `"finalizer" is optional, could be a function.`);
+
 	let error = undefined;
 	let index = 0;
 
@@ -43,6 +43,15 @@ export function* toConsumer(consumer, finalizer) {
 }
 
 /**
+ * @param {Function} consumer
+ * @param {Function} [finalizer]
+ * @returns {Observer}
+ */
+export function toConsumer(consumer, finalizer) {
+	return initializeObserver(consumerTarget(consumer, finalizer));
+}
+
+/**
  * @returns {Observer}
  */
 export function toNull() {
@@ -59,11 +68,9 @@ export function toNull() {
 	};
 }
 
-/**
- * @param {Function} setter
- * @returns {Observer}
- */
-export function* toValue(setter) {
+function* valueTarget(setter) {
+	assertFunctor(setter, `setter`);
+
 	let firstValue = undefined;
 	let count = 0;
 
@@ -75,10 +82,18 @@ export function* toValue(setter) {
 
 		assert(count === 1, `"toValue()" may be applied only to a sequence of one item, but got at least 2.`);
 	} finally {
-		assert(count === 0, `"toValue()" may be applied only to a sequence of one item, but got none.`);
+		assert(count > 0, `"toValue()" may be applied only to a sequence of one item, but got none.`);
 
 		if (count === 1) {
 			setter(firstValue);
 		}
 	}
+}
+
+/**
+ * @param {Function} setter
+ * @returns {Observer}
+ */
+export function toValue(setter) {
+	return initializeObserver(valueTarget(setter));
 }