+++ packages/undercut/src/utils.js
@@ -4,4 +4,5 @@ export { compare };
 export * from "./utils/function.js";
 export * from "./utils/iterable.js";
 export * from "./utils/language.js";
+export * from "./utils/promise.js";
 export { asc, desc } from "./utils/sorting.js";